requirements.txt
chromadb
pymupdf
openai
python-dotenv
openai
python-dotenv
python-telegram-bot
opik
pytest
